feat: update database schema for financial year support
- Add country_code, financial_year_type, start_month/day fields to teams table
- Update locations table country field to use 3-character ISO codes
- Enhance invoice numbering series for financial year reset tracking
- Update factories to use ISO country codes instead of full names
- Update seeders with proper country code format
- Maintain backward compatibility with existing data structure

+++ database/migrations/2025_07_05_200240_create_invoice_numbering_series_table.php
@@ -18,18 +18,18 @@ return new class extends Migration
             $table->string('name', 100); // e.g., "Default", "Dubai Branch", "Main Office"
             $table->string('prefix', 20); // e.g., "INV", "EST", "INV-DXB"
             $table->string('format_pattern', 255); // e.g., "{PREFIX}-{YEAR}-{MONTH}-{SEQUENCE:4}"
-            $table->integer('current_number'); // Current sequence number
+            $table->unsignedBigInteger('current_number'); // Current sequence number
             $table->string('reset_frequency', 20);
             $table->boolean('is_active');
             $table->boolean('is_default');
             $table->timestamp('last_reset_at')->nullable();
             $table->timestamps();
-            
+
             // Indexes for performance
             $table->index(['organization_id', 'location_id']);
             $table->index(['organization_id', 'is_active']);
             $table->index(['organization_id', 'is_default']);
-            
+
             // Note: Unique constraint for default series will be handled at application level
         });
     }
